+++ CHANGE_LOG.md
@@ -2,5 +2,4 @@
 
 ### Please report if you found the issues!
 
-* Display reload button only on resource pack selection screen
-* Open parent screen (Pause screen) when close advancement screen
+* Fix close advancements screen does not open previous screen properly